@@ -1,12 +1,6 @@
 <script lang="ts">
-	import { Toaster } from '$components/ui/sonner';
-	import { ModeWatcher } from 'mode-watcher';
-	import { inject } from '@vercel/analytics';
 	import '../app.css';
-
-	inject();
+	let { children } = $props();
 </script>
 
-<ModeWatcher />
-<Toaster />
-<slot />
+{@render children()}
